Seperating platform specific logic
