+++ Navigator.tsx
@@ -20,38 +20,11 @@ interface IProps {
 
 const Stack = createNativeStackNavigator();
 export const navigationRef = createNavigationContainerRef();
-export const onRemoteNotification = (notification: PushNotification) => {
-  const isClicked = notification.getData().userInteraction === 1;
-
-  if (isClicked) {
-    // Navigate user to details screen
-    if (navigationRef.isReady()) {
-      navigationRef.dispatch(
-        CommonActions.navigate(routes.details, {
-          title: notification.getMessage(),
-          content: notification.getTitle(),
-          sentAt: Math.round(Date.now() / 1000) - 1,
-        }),
-      );
-    }
-  }
-  // Use the appropriate result based on what you needed to do for this notification
-  const result = PushNotificationIOS.FetchResult.NoData;
-  notification.finish(result);
-};
 
 export default function Navigator(props: IProps) {
   const [credentials] = useCredentials();
   const isLoading = typeof credentials === 'undefined';
   const isSignedIn = !!credentials;
-
-  useEffect(() => {
-    const type = 'localNotification';
-    PushNotificationIOS.addEventListener(type, onRemoteNotification);
-    return () => {
-      PushNotificationIOS.removeEventListener(type);
-    };
-  }, []);
 
   return (
     <NavigationContainer ref={navigationRef}>
