+++ hooks/usePushNotificationHandler.tsx
@@ -1,0 +1,1 @@
+export default function usePushNotificationHandler() {}
