+++ screens/Home.tsx
@@ -5,10 +5,13 @@ import {styles} from '../constants';
 import {useCredentials} from '../hooks/useAuth';
 import {useNotifications} from '@magicbell/react-headless';
 import Notification from '../components/Notification';
+import usePushNotificationHandler from '../hooks/usePushNotificationHandler';
 
 export default function HomeScreen(): React.JSX.Element {
   const [_, __, logout] = useCredentials();
   const store = useNotifications();
+
+  usePushNotificationHandler()
 
   return (
     <SafeAreaView style={styles.sectionContainer}>
